Add files via upload
Created 1 enemy that walks back and forth, created an item that can be collected, and once collected, gives you time and disappears temporarily. If you touch an enemy, time is subtracted ,you are knocked back, the character sprite will flash, and the character will be invincible for a short time, meaning time will not be subtracted if you are hit while the sprite is flashing.